Refactored control structures
