+++ joern-cli/frontends/rubysrc2cpg/src/main/scala/io/joern/rubysrc2cpg/astcreation/AstForStatementsCreator.scala
@@ -5,24 +5,17 @@ import io.joern.rubysrc2cpg.datastructures.BlockScope
 import io.joern.rubysrc2cpg.passes.Defines
 import io.joern.rubysrc2cpg.passes.Defines.getBuiltInType
 import io.joern.x2cpg.{Ast, ValidationMode}
-import io.shiftleft.codepropertygraph.generated.{ControlStructureTypes, ModifierTypes}
+import io.shiftleft.codepropertygraph.generated.{ControlStructureTypes, DispatchTypes, ModifierTypes, Operators}
 import io.shiftleft.codepropertygraph.generated.nodes.{NewControlStructure, NewMethod, NewMethodRef, NewTypeDecl}
 
 trait AstForStatementsCreator(implicit withSchemaValidation: ValidationMode) { this: AstCreator =>
 
   protected def astsForStatement(node: RubyNode): Seq[Ast] = {
     baseAstCache.clear() // A safe approximation on where to reset the cache
-    node match
-      case node: WhileExpression            => astForWhileStatement(node) :: Nil
-      case node: DoWhileExpression          => astForDoWhileStatement(node) :: Nil
-      case node: UntilExpression            => astForUntilStatement(node) :: Nil
-      case node: IfExpression               => astForIfStatement(node) :: Nil
-      case node: UnlessExpression           => astForUnlessStatement(node) :: Nil
-      case node: ForExpression              => astForForExpression(node) :: Nil
+    node match {
+      case node: IfExpression               => astForIfStatement(node)
       case node: CaseExpression             => astsForCaseExpression(node)
       case node: StatementList              => astForStatementList(node) :: Nil
-      case node: SimpleCallWithBlock        => astForCallWithBlock(node) :: Nil
-      case node: MemberCallWithBlock        => astForCallWithBlock(node) :: Nil
       case node: ReturnExpression           => astForReturnStatement(node) :: Nil
       case node: AnonymousTypeDeclaration   => astForAnonymousTypeDeclaration(node) :: Nil
       case node: TypeDeclaration            => astForClassDeclaration(node)
@@ -34,33 +27,16 @@ trait AstForStatementsCreator(implicit withSchemaValidation: ValidationMode) { t
       case node: BreakStatement             => astForBreakStatement(node) :: Nil
       case node: SingletonStatementList     => astForSingletonStatementList(node)
       case _                                => astForExpression(node) :: Nil
+    }
   }
 
-  private def astForWhileStatement(node: WhileExpression): Ast = {
-    val conditionAst = astForExpression(node.condition)
-    val bodyAsts     = astsForStatement(node.body)
-    whileAst(Some(conditionAst), bodyAsts, Option(code(node)), line(node), column(node))
-  }
-
-  private def astForDoWhileStatement(node: DoWhileExpression): Ast = {
-    val conditionAst = astForExpression(node.condition)
-    val bodyAsts     = astsForStatement(node.body)
-    doWhileAst(Some(conditionAst), bodyAsts, Option(code(node)), line(node), column(node))
-  }
-
-  // `until T do B` is lowered as `while !T do B`
-  private def astForUntilStatement(node: UntilExpression): Ast = {
-    val notCondition = astForExpression(UnaryExpression("!", node.condition)(node.condition.span))
-    val bodyAsts     = astsForStatement(node.body)
-    whileAst(Some(notCondition), bodyAsts, Option(code(node)), line(node), column(node))
-  }
-
-  private def astForIfStatement(node: IfExpression): Ast = {
+  private def astForIfStatement(node: IfExpression): Seq[Ast] = {
     def builder(node: IfExpression, conditionAst: Ast, thenAst: Ast, elseAsts: List[Ast]): Ast = {
       val ifNode = controlStructureNode(node, ControlStructureTypes.IF, code(node))
       controlStructureAst(ifNode, Some(conditionAst), thenAst :: elseAsts)
     }
-    foldIfExpression(builder)(node)
+
+    foldIfExpression(builder)(node) :: Nil
   }
 
   /** Registers the currently set access modifier for the current type (until it is reset later).
@@ -106,126 +82,12 @@ trait AstForStatementsCreator(implicit withSchemaValidation: ValidationMode) { t
             Nil
   }
 
-  private def astForElseClause(node: RubyNode): Ast = {
-    node match
-      case elseNode: ElseClause =>
-        elseNode.thenClause match
-          case stmtList: StatementList => astForStatementList(stmtList)
-          case node =>
-            logger.warn(s"Expecting statement list in ${code(node)} ($relativeFileName), skipping")
-            astForUnknown(node)
-      case elseNode =>
-        logger.warn(s"Expecting else clause in ${code(elseNode)} ($relativeFileName), skipping")
-        astForUnknown(elseNode)
-  }
-
-  // `unless T do B` is lowered as `if !T then B`
-  private def astForUnlessStatement(node: UnlessExpression): Ast = {
-    val notConditionAst = astForExpression(UnaryExpression("!", node.condition)(node.condition.span))
-    val thenAst = node.trueBranch match
-      case stmtList: StatementList => astForStatementList(stmtList)
-      case _                       => astForStatementList(StatementList(List(node.trueBranch))(node.trueBranch.span))
-    val elseAsts = node.falseBranch.map(astForElseClause).toList
-    val ifNode   = controlStructureNode(node, ControlStructureTypes.IF, code(node))
-    controlStructureAst(ifNode, Some(notConditionAst), thenAst :: elseAsts)
-  }
-
-  private def astForForExpression(node: ForExpression): Ast = {
-    val forEachNode  = controlStructureNode(node, ControlStructureTypes.FOR, code(node))
-    val doBodyAst    = astsForStatement(node.doBlock)
-    val iteratorNode = astForExpression(node.forVariable)
-    val iterableNode = astForExpression(node.iterableVariable)
-    Ast(forEachNode).withChild(iteratorNode).withChild(iterableNode).withChildren(doBodyAst)
-  }
-
-  protected def astsForCaseExpression(node: CaseExpression): Seq[Ast] = {
-    def goCase(expr: Option[SimpleIdentifier]): List[RubyNode] = {
-      val elseThenClause: Option[RubyNode] = node.elseClause.map(_.asInstanceOf[ElseClause].thenClause)
-      val whenClauses                      = node.whenClauses.map(_.asInstanceOf[WhenClause])
-      val ifElseChain = whenClauses.foldRight[Option[RubyNode]](elseThenClause) {
-        (whenClause: WhenClause, restClause: Option[RubyNode]) =>
-          // We translate multiple match expressions into an or expression.
-          //
-          // A single match expression is compared using `.===` to the case target expression if it is present
-          // otherwise it is treated as a conditional.
-          //
-          // There may be a splat as the last match expression,
-          // `case y when *x then c end` or
-          // `case when *x then c end`
-          // which is translated to `x.include? y` and `x.any?` conditions respectively
-
-          val conditions = whenClause.matchExpressions.map { mExpr =>
-            expr.map(e => BinaryExpression(mExpr, "===", e)(mExpr.span)).getOrElse(mExpr)
-          } ++ (whenClause.matchSplatExpression.iterator.flatMap {
-            case splat @ SplattingRubyNode(exprList) =>
-              expr
-                .map { e =>
-                  List(MemberCall(exprList, ".", "include?", List(e))(splat.span))
-                }
-                .getOrElse {
-                  List(MemberCall(exprList, ".", "any?", List())(splat.span))
-                }
-            case e =>
-              logger.warn(s"Unrecognised RubyNode (${e.getClass}) in case match splat expression")
-              List(Unknown()(e.span))
-          })
-          // There is always at least one match expression or a splat
-          // a splat will become an unknown in condition at the end
-          val condition = conditions.init.foldRight(conditions.last) { (cond, condAcc) =>
-            BinaryExpression(cond, "||", condAcc)(whenClause.span)
-          }
-          val conditional = IfExpression(
-            condition,
-            whenClause.thenClause.asStatementList,
-            List(),
-            restClause.map { els => ElseClause(els.asStatementList)(els.span) }
-          )(node.span)
-          Some(conditional)
-      }
-      ifElseChain.iterator.toList
-    }
-    def generatedNode: StatementList = node.expression
-      .map { e =>
-        val tmp = SimpleIdentifier(None)(e.span.spanStart(this.tmpGen.fresh))
-        StatementList(
-          List(SingleAssignment(tmp, "=", e)(e.span)) ++
-            goCase(Some(tmp))
-        )(node.span)
-      }
-      .getOrElse(StatementList(goCase(None))(node.span))
-    astsForStatement(generatedNode)
-  }
-
   protected def astForStatementList(node: StatementList): Ast = {
     val block = blockNode(node)
     scope.pushNewScope(BlockScope(block))
     val statementAsts = node.statements.flatMap(astsForStatement)
     scope.popScope()
     blockAst(block, statementAsts)
-  }
-
-  /* `foo(<args>) do <params> <stmts> end` is lowered as a METHOD node shaped like so:
-   * ```
-   * <method_ref> = def <lambda>0(<params>)
-   *   <stmts>
-   * end
-   * foo(<args>, <method_ref>)
-   * ```
-   */
-  protected def astForCallWithBlock[C <: RubyCall](node: RubyNode & RubyCallWithBlock[C]): Ast = {
-    val Seq(typeRef, _)  = astForDoBlock(node.block): @unchecked
-    val typeRefDummyNode = typeRef.root.map(DummyNode(_)(node.span)).toList
-
-    // Create call with argument referencing the MethodRef
-    val callWithLambdaArg = node.withoutBlock match {
-      case x: SimpleCall => astForSimpleCall(x.copy(arguments = x.arguments ++ typeRefDummyNode)(x.span))
-      case x: MemberCall => astForMemberCall(x.copy(arguments = x.arguments ++ typeRefDummyNode)(x.span))
-      case x =>
-        logger.warn(s"Unhandled call-with-block type ${code(x)}, creating anonymous method structures only")
-        Ast()
-    }
-
-    callWithLambdaArg
   }
 
   protected def astForDoBlock(block: Block & RubyNode): Seq[Ast] = {
